Announcementのqueryのドメインモデルを用意することでパラメータを拡張しやすいようにした
