+++ internal/domain/announcement_query.go
@@ -1,0 +1,5 @@
+package domain
+
+type AnnouncementQuery struct {
+	IsActive *bool
+}
