+++ internal/handler/announcement.go
@@ -4,7 +4,6 @@ import (
 	"net/http"
 
 	api "github.com/fun-dotto/announcement-api/generated"
-	"github.com/fun-dotto/announcement-api/internal/domain"
 	"github.com/fun-dotto/announcement-api/internal/service"
 	"github.com/gin-gonic/gin"
 )
@@ -18,7 +17,8 @@ func NewHandler(announcementService *service.AnnouncementService) *Handler {
 }
 
 func (h *Handler) AnnouncementsList(c *gin.Context, params api.AnnouncementsListParams) {
-	announcements, err := h.announcementService.GetAnnouncements(params.IsActive)
+	announcementQuery := toDomainAnnouncementQuery(params)
+	announcements, err := h.announcementService.GetAnnouncements(announcementQuery)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
@@ -30,14 +30,4 @@ func (h *Handler) AnnouncementsList(c *gin.Context, params api.AnnouncementsList
 	}
 
 	c.JSON(http.StatusOK, apiAnnouncements)
-}
-
-func toApiAnnouncement(announcement domain.Announcement) api.Announcement {
-	return api.Announcement{
-		Id:       announcement.ID,
-		Title:    announcement.Title,
-		Date:     announcement.Date,
-		Url:      announcement.URL,
-		IsActive: announcement.IsActive,
-	}
 }
