+++ internal/service/announcement.go
@@ -3,7 +3,7 @@ package service
 import "github.com/fun-dotto/announcement-api/internal/domain"
 
 type AnnouncementRepository interface {
-	GetAnnouncements(isActive *bool) ([]domain.Announcement, error)
+	GetAnnouncements(query domain.AnnouncementQuery) ([]domain.Announcement, error)
 }
 
 type AnnouncementService struct {
@@ -14,6 +14,6 @@ func NewAnnouncementService(announcementRepository AnnouncementRepository) *Anno
 	return &AnnouncementService{announcementRepository: announcementRepository}
 }
 
-func (s *AnnouncementService) GetAnnouncements(isActive *bool) ([]domain.Announcement, error) {
-	return s.announcementRepository.GetAnnouncements(isActive)
+func (s *AnnouncementService) GetAnnouncements(query domain.AnnouncementQuery) ([]domain.Announcement, error) {
+	return s.announcementRepository.GetAnnouncements(query)
 }
